Add reading recommendation
The original recommendation has out-of-sequence problems, and some of it (e.g., description of how POST requests are used) is misleading. The proposed additional reading is more detailed in most respects (but not all, e.g. status codes), and more accurate in some respects (e.g., POST requests). It also gives detailed instructions for use of curl and Chrome dev tools to examine HTTP traffic.