Changed for consistency and updated
Use of promises is not an added requirement. Only the use of a promise-supporting library is. If actual use of promises is an intended requirement, that requirement should be stated and elaborated, since it isn’t obvious (at least to me) how one would benefit from using promises in this context. The proposed change assumes it isn’t an intended requirement. I also propose updating the library requirement to permit using request-promise-native.