Corrected misordering of modules
The Adding State module includes the Secure Session Cookie exercise, which includes App1, which includes a requirement to “render an html page with a form that does a post”. Nothing has prepared the Learner to satisfy this requirement yet, but the skills for that are developed in what was the next module. There does not appear to be any inverse dependency. So it is appropriate to list these two modules in inverse order.