Revised specs to make them plausible.
Not clear why the app3 spec would require updating the cookie on each request, unless it is to renew maxAge, which isn’t even mentioned. So I’m guessing that the intent is to require updating it on each POST request only. Amended accordingly. Also not clear what it means to “Update the form to store the user's first name, last name, and color in the session cookie object”. Amended to make sense.