Removed superfluous requirement
The current version of express-session doesn't need cookie-parser to perform its function. In fact the documentation suggests strongly against it:

Since version 1.5.0, the cookie-parser middleware no longer needs to be used for this module to work. This module now directly reads and writes cookies on req/res. Using cookie-parser may result in issues if the secret is not the same between this module and cookie-parser.

I was able to build all other requirements of the spec without cookie-parser