Revised exercise to better respect Learners.
Reference to Mac calculator’s rules was a dead link. Those rules do not seem to be available, and, in any case, the application does not deserve to be cloned. Revised version aims to respect Learners’ intelligence, creativity, and ambition, giving them a taste of the fact that they can improve on even huge enterprises’ applications. Amount of work for all 3 stages was disproportional for 1 exercise of 1 module in Phase 2, so Stage 3 has been defined as a stretch goal.