Correct obsolete column name
This is throwing an error on every attempt to record a user’s addition of a skill to the user’s repertoire by checking the skill’s checkbox. New checks persist only as long as the page remains loaded, but disappear thereafter, because not inserted into the skill_checks table.